@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.4.1+5
+version: 1.4.2+6
 
 environment:
-  sdk: ">=2.17.6 <3.0.0"
+  sdk: '>=3.3.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,7 +34,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  math_expressions: ^2.3.1
+  math_expressions: ^2.4.0
   flutter_bloc: ^8.1.2
   bloc: ^8.1.1
   adivery: ^4.2.0 
@@ -42,6 +42,14 @@ dependencies:
   percent_indicator: ^4.0.1
   age_calculator: ^1.0.0
   auto_size_text: ^3.0.0
+  units_converter: ^3.0.0
+  equatable: ^2.0.5
+  gap: ^3.0.1
+  flutter_svg: ^2.0.10+1
+  shared_preferences: ^2.2.2
+  package_info_plus: ^6.0.0
+  url_launcher: ^6.2.5
+  share_plus: ^8.0.2
 
 
 dev_dependencies:
@@ -70,6 +78,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/
+    - assets/svg/
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
