@@ -1,18 +1,15 @@
-import 'package:age_calculator/age_calculator.dart';
+import 'package:calculator/bloc/age/from_state.dart';
+import 'package:calculator/bloc/age/result_state.dart';
+import 'package:calculator/bloc/age/to_state.dart';
 
-abstract class IAgeState {}
-
-class AgeState extends IAgeState {
-  DateTime birthdate;
-  DateTime curentdate;
-  String minuts;
-  String hour;
-  String day;
-  String week;
-  String months;
-  String year;
-  DateDuration age;
-  DateDuration nextbirthday;
-  AgeState(this.curentdate, this.birthdate, this.minuts, this.hour, this.day,
-      this.week, this.months, this.year, this.age, this.nextbirthday);
+class AgeState {
+  final ToState to;
+  final FromState from;
+  final ResultState result;
+  const AgeState({required this.to, required this.from, required this.result});
+  AgeState copyWith(
+      {ToState? newto, FromState? newfrom, ResultState? newresult}) {
+    return AgeState(
+        to: newto ?? to, from: newfrom ?? from, result: newresult ?? result);
+  }
 }
