@@ -2,22 +2,23 @@ import '../../model/capabilities.dart';
 
 abstract class IUnitconversionEvent {}
 
-class Unitconversion extends IUnitconversionEvent {
-  String value;
-  Unitconversion(this.value);
+final class UnitconversionInitial extends IUnitconversionEvent {
+  final Capabilities from;
+  final Capabilities to;
+  UnitconversionInitial(this.from, this.to);
 }
 
-class Unit extends IUnitconversionEvent {
-  Capabilities unit;
-  Unit(this.unit);
+final class Unitconversion extends IUnitconversionEvent {
+  final String input;
+  Unitconversion(this.input);
 }
 
-class Unit1 extends IUnitconversionEvent {
-  Capabilities unit1;
-  Unit1(this.unit1);
+final class FromUnit extends IUnitconversionEvent {
+  final Capabilities from;
+  FromUnit(this.from);
 }
 
-class Settitel extends IUnitconversionEvent {
-  String titel;
-  Settitel(this.titel);
+final class ToUnit extends IUnitconversionEvent {
+  final Capabilities to;
+  ToUnit(this.to);
 }
