@@ -1,13 +1,28 @@
-import 'package:flutter/material.dart';
+import 'package:calculator/bloc/unitconversion/from_state.dart';
+import 'package:calculator/bloc/unitconversion/input_state.dart';
+import 'package:calculator/bloc/unitconversion/result_state.dart';
+import 'package:calculator/bloc/unitconversion/to_state.dart';
 
-import '../../model/capabilities.dart';
-
-abstract class IUnitconversionState {}
-
-class UnitconversionState extends IUnitconversionState {
-  String value;
-  Capabilities unit;
-  Capabilities unit1;
-  String result;
-  UnitconversionState(this.value, this.unit, this.unit1, this.result);
+class UnitconversionState {
+  final FromState from;
+  final ToState to;
+  final ResultState result;
+  final InputState input;
+  const UnitconversionState(
+      {required this.from,
+      required this.to,
+      required this.result,
+      required this.input});
+  UnitconversionState copyWith({
+    FromState? newfrom,
+    ToState? newto,
+    ResultState? newresult,
+    InputState? newinput,
+  }) {
+    return UnitconversionState(
+        from: newfrom ?? from,
+        to: newto ?? to,
+        result: newresult ?? result,
+        input: newinput ?? input);
+  }
 }
